Move UsageGraph to its own component
