+++ app/src/pages/logged-calls/index.tsx
@@ -18,26 +18,15 @@ import {
   Breadcrumb,
   BreadcrumbItem,
 } from "@chakra-ui/react";
-import {
-  LineChart,
-  Line,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip,
-  Legend,
-  ResponsiveContainer,
-} from "recharts";
 import { Ban, DollarSign, Hash } from "lucide-react";
-import { useMemo } from "react";
 
 import AppShell from "~/components/nav/AppShell";
 import PageHeaderContainer from "~/components/nav/PageHeaderContainer";
 import ProjectBreadcrumbContents from "~/components/nav/ProjectBreadcrumbContents";
 import { useSelectedProject } from "~/utils/hooks";
-import dayjs from "~/utils/dayjs";
 import { api } from "~/utils/api";
 import LoggedCallTable from "~/components/dashboard/LoggedCallTable";
+import UsageGraph from "~/components/dashboard/UsageGraph";
 
 export default function LoggedCalls() {
   const { data: selectedProject } = useSelectedProject();
@@ -46,16 +35,6 @@ export default function LoggedCalls() {
     { projectId: selectedProject?.id ?? "" },
     { enabled: !!selectedProject },
   );
-
-  const data = useMemo(() => {
-    return (
-      stats.data?.periods.map(({ period, numQueries, totalCost }) => ({
-        period,
-        Requests: numQueries,
-        "Total Spent (USD)": parseFloat(totalCost.toString()),
-      })) || []
-    );
-  }, [stats.data]);
 
   return (
     <AppShell requireAuth>
@@ -83,39 +62,7 @@ export default function LoggedCalls() {
                 </Heading>
               </CardHeader>
               <CardBody>
-                <ResponsiveContainer width="100%" height={400}>
-                  <LineChart data={data} margin={{ top: 5, right: 20, left: 10, bottom: 5 }}>
-                    <XAxis
-                      dataKey="period"
-                      tickFormatter={(str: string) => dayjs(str).format("MMM D")}
-                    />
-                    <YAxis yAxisId="left" dataKey="Requests" orientation="left" stroke="#8884d8" />
-                    <YAxis
-                      yAxisId="right"
-                      dataKey="Total Spent (USD)"
-                      orientation="right"
-                      unit="$"
-                      stroke="#82ca9d"
-                    />
-                    <Tooltip />
-                    <Legend />
-                    <CartesianGrid stroke="#f5f5f5" />
-                    <Line
-                      dataKey="Requests"
-                      stroke="#8884d8"
-                      yAxisId="left"
-                      dot={false}
-                      strokeWidth={2}
-                    />
-                    <Line
-                      dataKey="Total Spent (USD)"
-                      stroke="#82ca9d"
-                      yAxisId="right"
-                      dot={false}
-                      strokeWidth={2}
-                    />
-                  </LineChart>
-                </ResponsiveContainer>
+                <UsageGraph />
               </CardBody>
             </Card>
             <VStack spacing="4" width="300px" align="stretch">
@@ -127,7 +74,7 @@ export default function LoggedCalls() {
                       <Icon as={DollarSign} boxSize={4} color="gray.500" />
                     </HStack>
                     <StatNumber>
-                      ${parseFloat(stats.data?.totals?.totalCost?.toString() ?? "0").toFixed(2)}
+                      ${parseFloat(stats.data?.totals?.totalCost?.toString() ?? "0").toFixed(3)}
                     </StatNumber>
                   </Stat>
                 </CardBody>
